refactor(utils): reimplement file loading utilities using `pathlib`
