+++ src/declusor/util/__init__.py
@@ -1,4 +1,4 @@
-from .file import load_file, load_file_safely, load_library, load_payload
+from .filesystem import load_file, load_file_safely, load_library, load_payload
 from .format import convert_bytes_to_hex, format_bash_function_call, format_client_bash_code
 from .io import read_message, write_binary_message, write_error_message, write_message, write_warning_message
 from .parse import parse_command_arguments
